🧬 Move the modules about command line features into new sub-package *pymock_api.command*.
🧬 Refactor the test implementation code usage align with migration.
