+++ pymock_api/command/cmd.py
@@ -20,7 +20,7 @@ from collections import namedtuple
 from dataclasses import dataclass
 from typing import Any, Dict, List, Optional, Tuple
 
-from .__pkg_info__ import __version__
+from ..__pkg_info__ import __version__
 
 SUBCOMMAND: List[str] = []
 COMMAND_OPTIONS: List["MetaCommandOption"] = []
